+++ packages/react-magma-dom/src/components/DatePicker/DatePicker.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { DatePicker } from '.';
+import { Card } from '../Card';
 
 export default {
   component: DatePicker,
@@ -8,4 +9,8 @@ export default {
 
 export const Default = () => {
   return <DatePicker labelText="Date" />;
+};
+
+export const Inverse = () => {
+  return <Card isInverse><DatePicker labelText="Date" isInverse /></Card>;
 };