+++ packages/react-magma-dom/src/components/Heading/index.tsx
@@ -20,7 +20,7 @@ export interface HeadingProps extends TypographyProps<HTMLHeadingElement> {
   color?: TypographyColor;
   /**
    * Additional styles for typography based on the context of the content
-   * @default TypographyColor.default
+   * @default TypographyContextVariant.default
    */
   contextVariant?: TypographyContextVariant;
   isInverse?: boolean;
@@ -37,7 +37,7 @@ export interface HeadingProps extends TypographyProps<HTMLHeadingElement> {
 
 export const Heading = React.forwardRef<HTMLHeadingElement, HeadingProps>(
   (props, ref) => {
-    const { level, testId, visualStyle, children, ...other } = props;
+    const { level, testId, visualStyle, children, contextVariant, ...other } = props;
     const theme = React.useContext(ThemeContext);
 
     const stylesFromLevel = {
@@ -55,6 +55,7 @@ export const Heading = React.forwardRef<HTMLHeadingElement, HeadingProps>(
       <TypographyComponent
         {...other}
         as={headingElement}
+        contextVariant={contextVariant}
         data-testid={testId}
         isInverse={useIsInverse(props.isInverse)}
         ref={ref}