+++ website/react-magma-docs/src/pages/index.js
@@ -13,8 +13,8 @@ const IndexPage = () => (
           styles={css`
             @import url('https://fonts.googleapis.com/css?family=Abel&display=swap');
             .content {
-              background: ${theme.colors.foundation};
-              color: ${theme.colors.neutral};
+              background: ${theme.colors.neutral100};
+              color: ${theme.colors.neutral700};
               padding: 0;
               overflow: hidden;
             }